refactor:add images dark and light in public;changed the path to images in Theme
