+++ src/Page_404/Page_404.test.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import Page_404 from './Page_404';
+import Page_404 from '../pages/404';
 import '@testing-library/jest-dom';
 import { BrowserRouter } from 'react-router-dom';
