+++ src/components/Download/Download.tsx
@@ -3,6 +3,7 @@ import { RootState } from '../../redux/store';
 import { PeopleArray } from '../../types/types';
 import styles from './Download.module.css';
 
+
 export function convertToCSV(data: PeopleArray[]) {
   const csvHeaders = [
     'name      ',
@@ -44,14 +45,14 @@ function Download() {
   const url = window.URL.createObjectURL(blob);
 
   return (
-    <a
+   <a
       download={filename}
       href={url}
       role="link"
       className={styles.flyout_container_download_button}
     >
-      Download
-    </a>
+        Download
+      </a>
   );
 }
 export default Download;
