+++ src/components/Seach/Seach.test.tsx
@@ -1,21 +1,21 @@
 import { render } from '@testing-library/react';
 import Seach from './Seach';
 import '@testing-library/jest-dom';
-//import { BrowserRouter } from 'react-router-dom';
+
 
 test('extracting value from local storage when component mounts', () => {
   const savedSearchLocal = JSON.stringify('C-3PO');
   localStorage.setItem('search', savedSearchLocal);
   const { getByPlaceholderText } = render(
-    <BrowserRouter>
-      {' '}
+    
+     
       <Seach
         enterHandler={function (): void {
           throw new Error('Function not implemented.');
         }}
         savedSearchLocal={''}
       />
-    </BrowserRouter>
+    
   );
 
   const searchInput = getByPlaceholderText('Enter the name of the person');
