+++ src/pages/404.tsx
@@ -3,7 +3,7 @@ import ErrorMessage from '../components/ErrorMessage/ErrorMessage';
 function Page_404() {
   return (
     <div>
-      <div>Page_404</div>;
+      <div>Page_404</div>
       <ErrorMessage />
     </div>
   );
