+++ src/pages/_app.tsx
@@ -1,8 +1,7 @@
 import '../styles/global.css';
 import type { AppProps } from 'next/app';
 import { Provider } from 'react-redux';
-import { store} from '../redux/store';
-
+import { store } from '../redux/store';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
