+++ src/pages/index.tsx
@@ -1,6 +1,8 @@
-
 import Page from '../Page';
+import Themes from '../components/Themes/Themes';
+import ThemeProvider from '../context/ThemeProvider';
 
 export default function Home() {
-  return <Page />;
+  return (
+    <> <ThemeProvider><Themes /><Page /></ThemeProvider></>);
 }
